requirements.txt
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
coverage==6.3.3
Django==4.0.4
django-shell-plus==1.1.7
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.1
pytest-cov==3.0.0
pytest-django==4.5.2
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
